+++ mapie/classification.py
@@ -43,11 +43,16 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
         Method to choose for prediction interval estimates.
         Choose among:
 
+        - "naive", sum of the probabilities until the 1-alpha thresold.
         - "score", based on the the scores
           (i.e. 1 minus the softmax score of the true label)
           on the calibration set.
         - "cumulated_score", based on the sum of the softmax outputs of the
           labels until the true label is reached, on the calibration set.
+        - "top_k", based on the sorted index of the probability of the true
+        label in the softmax outputs, on the calibration set. In case two
+        probabilities are equal, both are taken, thus, the size of some p
+        prediction sets may be different from the others.
 
           By default "score".
 
@@ -148,7 +153,7 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
      [False False  True]]
     """
 
-    valid_methods_ = ["score", "cumulated_score"]
+    valid_methods_ = ["naive", "score", "cumulated_score", "top_k"]
 
     def __init__(
         self,
@@ -454,7 +459,7 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
                 (
                     y_proba_last_cumsumed[:, iq]
                     - quantile
-                ) / y_pred_proba_last[:, iq]
+                ) / np.squeeze(y_pred_proba_last[:, :, iq])
                 for iq, quantile in enumerate(self.quantiles_)
             ], axis=1,
         )
@@ -519,7 +524,10 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
         y_pred_proba = self.single_estimator_.predict_proba(X)
         y_pred_proba = self._check_proba_normalized(y_pred_proba)
         self.n_samples_val_ = X.shape[0]
-        if self.method == "score":
+
+        if self.method == "naive":
+            self.conformity_scores_ = np.empty(y_pred_proba.shape)
+        elif self.method == "score":
             self.conformity_scores_ = np.take_along_axis(
                 1 - y_pred_proba, y.reshape(-1, 1), axis=1
             )
@@ -542,6 +550,17 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
             random_state = check_random_state(self.random_state)
             u = random_state.uniform(size=len(y_pred_proba)).reshape(-1, 1)
             self.conformity_scores_ -= u*y_proba_true
+        elif self.method == "top_k":
+            # Here we reorder the labels by decreasing probability
+            # and get the position of each label from decreasing probability
+            index = np.argsort(
+                np.fliplr(np.argsort(y_pred_proba, axis=1))
+            )
+            self.conformity_scores_ = np.take_along_axis(
+                index,
+                y.reshape(-1, 1),
+                axis=1
+            )
 
         else:
             raise ValueError(
@@ -622,14 +641,19 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
         if alpha_ is None:
             return np.array(y_pred)
         else:
+
+            # Choice of the quantile
             check_alpha_and_n_samples(alpha_, n)
-            self.quantiles_ = np.stack([
-                np.quantile(
-                    self.conformity_scores_,
-                    ((n + 1) * (1 - _alpha)) / n,
-                    interpolation="higher"
-                ) for _alpha in alpha_
-            ])
+            if self.method == "naive":
+                self.quantiles_ = 1 - alpha_
+            else:
+                self.quantiles_ = np.stack([
+                    np.quantile(
+                        self.conformity_scores_,
+                        ((n + 1) * (1 - _alpha)) / n,
+                        interpolation="higher"
+                    ) for _alpha in alpha_
+                ])
             if self.method == "score":
                 prediction_sets = np.stack(
                     [
@@ -638,7 +662,7 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
                     ],
                     axis=2,
                 )
-            elif self.method == "cumulated_score":
+            elif self.method in ["cumulated_score", "naive"]:
                 # sort labels by decreasing probability
                 index_sorted = np.fliplr(np.argsort(y_pred_proba, axis=1))
                 # sort probabilities by decreasing order
@@ -663,24 +687,19 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
                 # get the probability of the last included label
                 y_pred_proba_last = np.stack(
                     [
-                        np.squeeze(
-                            np.take_along_axis(
-                                y_pred_proba,
-                                y_pred_index_last[:, iq].reshape(-1, 1),
-                                axis=1
-                            )
+                        np.take_along_axis(
+                            y_pred_proba,
+                            y_pred_index_last[:, iq].reshape(-1, 1),
+                            axis=1
                         )
                         for iq, _ in enumerate(self.quantiles_)
-                    ], axis=1
+                    ], axis=2
                 )
                 # get the prediction set by taking all probabilities above the
                 # last one
                 prediction_sets = np.stack(
                     [
-                        np.ma.masked_greater_equal(
-                            y_pred_proba,
-                            y_pred_proba_last[:, iq].reshape(-1, 1) - EPSILON
-                        ).mask
+                        y_pred_proba >= y_pred_proba_last[:, :, iq] - EPSILON
                         for iq, _ in enumerate(self.quantiles_)
                     ], axis=2
                 )
@@ -692,6 +711,30 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):  # type: ignore
                         y_pred_proba_cumsum,
                         y_pred_proba_last
                     )
+            elif self.method == "top_k":
+                index_sorted = np.fliplr(np.argsort(y_pred_proba, axis=1))
+                y_pred_index_last = np.stack(
+                    [
+                        index_sorted[:, quantile]
+                        for quantile in self.quantiles_
+                    ], axis=1
+                )
+                y_pred_proba_last = np.stack(
+                    [
+                        np.take_along_axis(
+                            y_pred_proba,
+                            y_pred_index_last[:, iq].reshape(-1, 1),
+                            axis=1
+                        )
+                        for iq, _ in enumerate(self.quantiles_)
+                    ], axis=2
+                )
+                prediction_sets = np.stack(
+                    [
+                        y_pred_proba >= y_pred_proba_last[:, :, iq] - EPSILON
+                        for iq, _ in enumerate(self.quantiles_)
+                    ], axis=2
+                )
             else:
                 raise ValueError(
                     "Invalid method. "