+++ mapie/tests/test_classification.py
@@ -21,7 +21,7 @@ from mapie._typing import ArrayLike
 
 
 METHODS = ["score", "cumulated_score"]
-WRONG_METHODS = ["scores", "cumulated", "test", ""]
+WRONG_METHODS = ["scores", "cumulated", "test", "nave", ""]
 WRONG_INCLUDE_LABELS = ["randomised", "True", "False", "", "other"]
 Y_PRED_PROBA = [
     np.array(
@@ -104,6 +104,26 @@ STRATEGIES = {
             include_last_label='randomized'
         )
     ),
+    "naive": (
+        Params(
+            method="naive",
+            cv="prefit",
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label=True
+        )
+    ),
+    "top_k": (
+        Params(
+            method="top_k",
+            cv="prefit",
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label=True
+        )
+    ),
 }
 
 COVERAGES = {
@@ -111,6 +131,8 @@ COVERAGES = {
     "cumulated_score_include": 1,
     "cumulated_score_not_include": 5/9,
     "cumulated_score_randomized": 8/9,
+    "naive": 1,
+    "top_k": 1
 }
 
 y_toy_mapie = {
@@ -155,6 +177,28 @@ y_toy_mapie = {
         [True, True, False],
         [False, True, True],
         [False, False, True],
+        [False, True, True],
+        [False, True, True],
+    ],
+    "naive": [
+        [True, False, False],
+        [True, False, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, True],
+        [False, True, True],
+        [False, True, True],
+        [False, False, True],
+        [False, False, True],
+    ],
+    "top_k": [
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, True],
+        [False, True, True],
+        [False, True, True],
         [False, True, True],
         [False, True, True],
     ],